API: Implement Laundry Booking users
Two new models are added to interact with the user collection of the
Laundry Booking DB. One model to upsert a user and another to read and
modify user bookings (the model is there but need to implement the part
to actually modify bookings by adding/updating/removing bookings).

Also a new method in the laundry booking router to add new users. These
users should already be added as auth users with username, email, and
password. Now they will be added to the Laundry Booking 'user'
collection with data related to the laundry booking, not auth. So, same
user, different data in different collections.