API: Refactor/simplify moving models to mongob
MongoDB models were implemented in each package (auth, laundry_booking),
but moving in that direction is nested, and flat is better. So the
models are moved to mongodb, making all the structure simpler and
easier.

Bad side: I lose some beautiful and elegant solution to provide pymongo
support in the auth package, but the extra complexity makes that not
worth it except for bragging maybe.

Other refactorings:
- Change typed dicts for pydantic models, should be better.
- Use __init__.py files as barrel files to simplify all the imports from
the local packages.
- Change docstrings accordingly.
- Finish refactoring imports (the main local imports are now made from
`src` to clearly identify local imports, except for subpackages where
imports are made relative to the subpackage to make it easier to move
these subpackages from project to project if needed).
- Change dockerfile accordingly (the full `api` folder needs to be
copied into the container instead of only the `src` folder if imports
are to be resolved correctly).